requirements.txt
protobuf==3.5.1
grpcio==1.8.2
grpcio-tools==1.8.2
